New animated job (#14)
- Implementation of a new dual animated job for Homer and Bart called
"Strangle Bart". It requires code editing for become available. Credits
to Discord user @zaloran for designing the animations.

- Premium Scratch-Rs are back in Kwik-E-Mart.

- New decoration: Box of Donuts. It has the special ability to change
the landing point of Springfield. Credits to Discord user @zaloran for
designing it.

- Fix the job "DanceTheNightAway" from Marge and Homer which was not
playing the animations correctly.

- Change all ids related to the NewHorizons DLC. As a side effect,
everyone will have to buy Moe's Apartment and Zoltar Machine again after
pushing this to the server.

---------

Co-authored-by: Dractiums <dractiums@gmail.com>